pyproject.toml
[tool.poetry]
name = "nice-fetch-ai-adapter"
version = "0.1.0"
description = ""
authors = ["Helena Adamkova <58051865+Ellenn-A@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.111.0"
uvicorn = "~0.20.0"
pytest-mock = "^3.14.0"
uagents = "^0.11.1"
mock = "^5.1.0"
requests-mock = "^1.12.1"
flake8 = "^7.0.0"
httpx = "^0.27.0"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.23.6"
respx = "^0.21.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
